# packages/forge-connectivity-service/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Forge Connectivity Service",
    "description": "Content-blind presence, mailbox, and key-package transport for independently operated Forge peers.",
    "version": "0.1.0",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "components": {
    "securitySchemes": {
      "ForgeChannelSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Ed25519 possession proof: ForgeChannel v1.<SPKI>.<unix-seconds>.<nonce>.<signature>. It is unrelated to Forge bearer tokens."
      }
    },
...
  },
  "paths": {
    "/.well-known/forge-connectivity": {
      "get": {
        "operationId": "getForgeConnectivityDiscovery",
        "summary": "Discover the Forge connectivity service contract and limits.",
        "tags": [
          "Discovery"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "protocol",
                    "service",
                    "version",
                    "auth",
                    "capabilities",
                    "limits",
                    "routes"
                  ],
                  "properties": {
                    "protocol": {
                      "type": "string",
                      "enum": [
                        "forge-connectivity/1"
                      ]
                    },
                    "service": {
                      "type": "string",
                      "enum": [
                        "forge-connectivity-service"
                      ]
                    },
                    "version": {
                      "type": "string"
                    },
                    "auth": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "scheme",
                        "algorithm",
                        "channelDerivation",
                        "replayProtection"
                      ],
                      "properties": {
                        "scheme": {
                          "type": "string",
                          "enum": [
                            "ForgeChannel"
                          ]
                        },
                        "algorithm": {
                          "type": "string",
                          "enum": [
                            "Ed25519"
                          ]
                        },
                        "channelDerivation": {
                          "type": "string",
                          "enum": [
                            "sha256-spki-v1"
                          ]
                        },
                        "replayProtection": {
                          "type": "string",
                          "enum": [
                            "timestamp-nonce-v1"
                          ]
                        }
                      }
                    },
                    "capabilities": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "presence",
                        "envelopes",
                        "keyPackages",
                        "longPoll",
                        "ciphertextOnly"
                      ],
                      "properties": {
                        "presence": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "envelopes": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "keyPackages": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "longPoll": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        },
                        "ciphertextOnly": {
                          "type": "boolean",
                          "enum": [
                            true
                          ]
                        }
                      }
                    },
                    "limits": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "maxEnvelopeBytes",
                        "maxPresenceBytes",
                        "maxKeyPackageBytes",
                        "maxChannelEnvelopeCount",
                        "maxChannelEnvelopeBytes",
                        "maxChannelRetainedEnvelopeCount",
                        "maxChannelKeyPackageCount",
                        "maxChannelKeyPackageBytes",
                        "channelRequestsPerMinute",
                        "channelBurstRequests",
                        "maxCursorPageSize",
                        "maxLongPollSeconds",
                        "maxEnvelopeTtlSeconds",
                        "maxPresenceTtlSeconds",
                        "maxKeyPackageTtlSeconds"
                      ],
                      "properties": {
                        "maxEnvelopeBytes": {
                          "type": "integer"
                        },
                        "maxPresenceBytes": {
                          "type": "integer"
                        },
                        "maxKeyPackageBytes": {
                          "type": "integer"
                        },
                        "maxChannelEnvelopeCount": {
                          "type": "integer"
                        },
                        "maxChannelEnvelopeBytes": {
                          "type": "integer"
                        },
                        "maxChannelRetainedEnvelopeCount": {
                          "type": "integer"
                        },
                        "maxChannelKeyPackageCount": {
                          "type": "integer"
                        },
                        "maxChannelKeyPackageBytes": {
                          "type": "integer"
                        },
                        "channelRequestsPerMinute": {
                          "type": "integer"
                        },
                        "channelBurstRequests": {
                          "type": "integer"
                        },
                        "maxCursorPageSize": {
                          "type": "integer"
                        },
                        "maxLongPollSeconds": {
                          "type": "integer"
                        },
                        "maxEnvelopeTtlSeconds": {
                          "type": "integer"
                        },
                        "maxPresenceTtlSeconds": {
                          "type": "integer"
                        },
                        "maxKeyPackageTtlSeconds": {
                          "type": "integer"
                        }
                      }
                    },
                    "routes": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "presence",
                        "envelopes",
                        "envelopeAck",
                        "keyPackages",
                        "health"
                      ],
                      "properties": {
                        "presence": {
                          "type": "string"
                        },
                        "envelopes": {
                          "type": "string"
                        },
                        "envelopeAck": {
                          "type": "string"
                        },
                        "keyPackages": {
                          "type": "string"
                        },
                        "health": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "AMBIGUOUS_REQUEST_HEADERS",
                            "AUTH_INVALID",
                            "AUTH_REPLAYED",
                            "AUTH_STALE",
                            "BLOB_INVALID",
                            "BLOB_TOO_LARGE",
                            "CONTENT_ENCODING_UNSUPPORTED",
                            "CONTENT_TYPE_UNSUPPORTED",
                            "CREDENTIAL_NOT_ALLOWED",
                            "CURSOR_INVALID",
                            "IDEMPOTENCY_CONFLICT",
                            "IDEMPOTENCY_REQUIRED",
                            "INTERNAL_ERROR",
                            "NOT_FOUND",
                            "POLL_LIMIT_EXCEEDED",
                            "QUOTA_EXCEEDED",
                            "RATE_LIMITED",
                            "REPLAY_CONFLICT",
                            "REQUEST_BODY_NOT_ALLOWED",
                            "SENSITIVE_METADATA_NOT_ALLOWED",
                            "SERVICE_CLOSING",
                            "STORAGE_UNAVAILABLE",
                            "VALIDATION_ERROR"
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}